sota: VarPro per-batch solve -> fp32 (4090 cripples fp64 ~64x) — the speed unlock
The convergence test showed VarPro reaches low ESR in ~5x FEWER epochs than joint (the
always-optimal readout trains the trunk against a perfect head from step 1), but the
fp64 per-batch solve was ~6x slower/epoch and cancelled the win. Root cause: the RTX 4090
cripples fp64 to ~1/64 of fp32. Switching the per-batch differentiable solve to fp32
(+ scale-relative ridge for conditioning): 18.5 ms/batch vs joint's ~16 ms (was ~6x) and
stable. So the 5x epoch speedup now becomes a real ~5x wall-clock training speedup at
equal accuracy. Re-running varpro_conv to demonstrate.

+++ experiments/sota/varpro_probe.py
@@ -162,12 +162,14 @@ def _solve_diff(f: torch.Tensor, y: torch.Tensor, ridge: float, warmup: int
     bsz, feat, _ = f.shape
     ones = torch.ones(bsz, 1, f.shape[2], device=f.device)
     fb = torch.cat([f, ones], 1)[:, :, warmup:]  # (B, F+1, Tw)
-    hcols = fb.permute(1, 0, 2).reshape(feat + 1, -1).double()  # (F+1, B*Tw), fp64
-    yt = y[:, warmup:].reshape(-1).double()  # (B*Tw,)
+    # fp32: the 4090 cripples fp64 to ~1/64 of fp32, so the per-batch solve was the
+    # whole VarPro slowdown. fp32 + scale-relative ridge keeps it well-conditioned.
+    hcols = fb.permute(1, 0, 2).reshape(feat + 1, -1)  # (F+1, B*Tw)
+    yt = y[:, warmup:].reshape(-1)  # (B*Tw,)
     a = hcols @ hcols.T
-    a += ridge * a.diag().mean() * torch.eye(feat + 1, device=f.device, dtype=torch.float64)
+    a += ridge * a.diag().mean() * torch.eye(feat + 1, device=f.device)
     w = torch.linalg.solve(a, hcols @ yt)  # (F+1,)
-    return (w @ hcols).reshape(bsz, -1).float()  # (B, Tw)
+    return (w @ hcols).reshape(bsz, -1)  # (B, Tw)
 
 
 def train(net: VarProNet, segs, device: str, varpro: bool, log) -> None:
